initial linux commit
* Add linux capture type.
* Unifies linux and BSD variants as much as possible.
* Adds Vagrant box for testing purposes.
* Adds Pry as development dependency.
* Reformats capture types into their own module.
