+++ .gitignore
@@ -1,4 +1,5 @@
 /.bundle/
+/.vagrant/
 /.yardoc
 /_yardoc/
 /coverage/
@@ -9,3 +10,4 @@
 
 # rspec failure tracking
 .rspec_status
+ubuntu-xenial-16.04-cloudimg-console.log
